Refactor: 엔티티 타입 변경 및 디폴트 값 설정
- tinyint에 맞는 byte 타입으로 변경
- prepersist 어노테이션을 활용해 디폴트값 설정
- 생성일 로직 변경

+++ src/main/java/com/tanppoppo/testore/testore/exam/entity/ExamItemEntity.java
@@ -31,7 +31,7 @@ public class ExamItemEntity {
 
     @NotNull
     @Column(name = "answer", columnDefinition = "TINYINT")
-    private Integer answer;
+    private Byte answer;
 
     @Column(name = "paragraph_description")
     private String paragraphDescription;
